Updates with a working 5 minute countdown timer on the a1-4 outputs. The timer is set by uploading a dim level.
Need to change this to be a 10 min countdown as I think that this would
be the shortes time that you would want.

A1-4 are also independent of each other and will count down inividually.

They are also set to not require any interaction once set and so will
turn off by them selves.